Fixed the scaling and anti-aliasing issues

+++ src/Glyph.cpp
@@ -3,10 +3,10 @@
 Glyph::Glyph(const char& character) {
 	this->texture = Texture(std::to_string(character), false);
 	this->character = character;
-	this->size = glm::ivec2(0, 0);
-	this->advance = glm::ivec2(0, 0);
-	this->horizontalBearing = glm::ivec2(0, 0);
-	this->verticalBearing = glm::ivec2(0, 0);
+	this->size = glm::vec2(0, 0);
+	this->advance = glm::vec2(0, 0);
+	this->horizontalBearing = glm::vec2(0, 0);
+	this->verticalBearing = glm::vec2(0, 0);
 }
 Glyph::~Glyph() {
